+++ ErsatzTV.Core/Scheduling/PlayoutBuildWarnings.cs
@@ -1,0 +1,15 @@
+namespace ErsatzTV.Core.Scheduling;
+
+public class PlayoutBuildWarnings
+{
+    public void Merge(PlayoutBuildWarnings warnings)
+    {
+        TailFillerTooLong += warnings.TailFillerTooLong;
+        MidRollContentWithoutChapters += warnings.MidRollContentWithoutChapters;
+        DurationFillerSkipped += warnings.DurationFillerSkipped;
+    }
+
+    public int TailFillerTooLong { get; set; }
+    public int MidRollContentWithoutChapters { get; set; }
+    public int DurationFillerSkipped { get; set; }
+}